Tidy diagnostic log details
Agent-Logs-Url: https://github.com/medievil1/main/sessions/675eaaf4-d451-40bb-9096-0b7a43cccfde

+++ src/devices/audio/AudioDiagnostics.h
@@ -18,7 +18,7 @@ inline uint32_t PeakAbsoluteSampleAmplitude(const int16_t* samples, size_t sampl
 	for (size_t i = 0; i < sampleCount; ++i) {
 		const int32_t signedSample = static_cast<int32_t>(samples[i]);
 		const uint32_t magnitude = signedSample < 0
-			? static_cast<uint32_t>(-static_cast<int64_t>(signedSample))
+			? static_cast<uint32_t>(-signedSample)
 			: static_cast<uint32_t>(signedSample);
 		peak = std::max(peak, magnitude);
 	}
